+++ Makie/src/basic_recipes/band.jl
@@ -1,11 +1,17 @@
 """
-    band(x, ylower, yupper; kwargs...)
-    band(lower, upper; kwargs...)
-    band(x, lowerupper; kwargs...)
+Plots a band between lower and upper bounds.
 
-Plots a band from `ylower` to `yupper` along `x`. The form `band(lower, upper)` plots a [ruled surface](https://en.wikipedia.org/wiki/Ruled_surface)
-between the points in `lower` and `upper`.
-Both bounds can be passed together as `lowerupper`, a vector of intervals.
+## Arguments
+
+* `xs, ys_lower, ys_upper` Plots a band from `ys_lower` to `ys_upper` along `xs`. `xs` is an
+    `AbstractVector{<:Real}` containing x-values, and `ys_lower`, `ys_upper` are
+    `AbstractVector{<:Real}` containing the lower and upper y-limits of the band.
+* `xs, lowerupper` Plots a band where `lowerupper` is an `AbstractVector{<:Interval}` containing the
+    lower and upper limits of the band as intervals.
+* `lower, upper` Plots a [ruled surface](https://en.wikipedia.org/wiki/Ruled_surface) between the
+    points in `lower` and `upper`. Both are `AbstractVector{<:Point{D, <:Real}}` containing the
+    (x, y) or (x, y, z) coordinates of the lower and upper limits of the band respectively.
+* Setting `direction = :y` will reinterpret (x, y) as (y, x). (2D only)
 """
 @recipe Band (lowerpoints, upperpoints) begin
     documented_attributes(Mesh)...
@@ -24,8 +30,9 @@ function convert_arguments(::Type{<:Band}, x, ylower, yupper)
     return (Point2{float_type(x, ylower)}.(x, ylower), Point2{float_type(x, yupper)}.(x, yupper))
 end
 
-convert_arguments(P::Type{<:Band}, x::AbstractVector{<:Number}, y::AbstractVector{<:Interval}) =
-    convert_arguments(P, x, leftendpoint.(y), rightendpoint.(y))
+function convert_arguments(P::Type{<:Band}, x::AbstractVector{<:Number}, y::AbstractVector{<:Interval})
+    return convert_arguments(P, x, leftendpoint.(y), rightendpoint.(y))
+end
 
 function band_connect(n)
     ns = 1:(n - 1)
@@ -33,7 +40,7 @@ function band_connect(n)
     return [GLTriangleFace.(ns, ns .+ 1, ns2); GLTriangleFace.(ns .+ 1, ns2 .+ 1, ns2)]
 end
 
-function Makie.plot!(plot::Band)
+function plot!(plot::Band)
     nanpoint(::Type{<:Point3}) = Point3(NaN)
     nanpoint(::Type{<:Point2}) = Point2(NaN)
     map!(plot, [:lowerpoints, :upperpoints, :direction], :coordinates) do lowerpoints, upperpoints, direction
@@ -127,22 +134,5 @@ end
 
 export fill_between!
 
-function attribute_examples(::Type{Band})
-    return Dict(
-        :direction => [
-            Example(
-                code = """
-                fig = Figure()
-                location = range(0, 4pi, length = 200)
-                lower =   cos.(location) .- location
-                upper = .-cos.(location) .+ location .+ 5
-                band(fig[1, 1], location, lower, upper,
-                    axis = (; title = "direction = :x"))
-                band(fig[1, 2], location, lower, upper, direction = :y,
-                    axis = (; title = "direction = :y"))
-                fig
-                """
-            ),
-        ],
-    )
-end
+# attribute_examples for Band has been moved to documentation/plots/band.md
+# under the "## Attributes" section and is now loaded automatically.