+++ Makie/src/basic_recipes/hvlines.jl
@@ -1,10 +1,12 @@
 """
-    hlines(ys; xmin = 0.0, xmax = 1.0, attrs...)
+Draws horizontal lines across an `Axis`.
 
-Create horizontal lines across a `Scene` with 2D projection.
-The lines will be placed at `ys` in data coordinates and `xmin` to `xmax`
-in scene coordinates (0 to 1). All three of these can have single or multiple values because
-they are broadcast to calculate the final line segments.
+## Arguments
+
+* `ys` Positions where lines will be drawn in data coordinates. Can be a `Real` or
+    `AbstractVector{<:Real}`. The lines span from `xmin` to `xmax` in relative coordinates (0 to 1).
+* `xmin, xmax` Attributes for setting the start and end positions of lines in relative (0 .. 1) space.
+    Can be a `Real` or `AbstractVector{<:Real}`
 """
 @recipe HLines begin
     "The start of the lines in relative axis units (0 to 1) along the x dimension."
@@ -15,12 +17,14 @@ they are broadcast to calculate the final line segments.
 end
 
 """
-    vlines(xs; ymin = 0.0, ymax = 1.0, attrs...)
+Draws vertical lines across an `Axis`.
 
-Create vertical lines across a `Scene` with 2D projection.
-The lines will be placed at `xs` in data coordinates and `ymin` to `ymax`
-in scene coordinates (0 to 1). All three of these can have single or multiple values because
-they are broadcast to calculate the final line segments.
+## Arguments
+
+* `xs` Positions where lines will be drawn in data coordinates. Can be a `Real` or
+    `AbstractVector{<:Real}`. The lines span from `ymin` to `ymax` in relative coordinates (0 to 1).
+* `ymin, ymax` Attributes for setting the start and end positions of lines in relative (0 .. 1) space.
+    Can be a `Real` or `AbstractVector{<:Real}`
 """
 @recipe VLines begin
     "The start of the lines in relative axis units (0 to 1) along the y dimension."