+++ Makie/src/stats/violin.jl
@@ -1,12 +1,6 @@
 """
-    violin(x, y)
-
-Draws a violin plot which consists of pairs of density plots draw side by side.
+Draws a violin plot which consists of pairs of density plots drawn side by side.
 The density pairs can be sourced from the same or from different data.
-
-## Arguments
-- `x`: positions of the categories
-- `y`: variables whose density is computed
 """
 @recipe Violin (x, y) begin
     "Number of points used per density plot."
@@ -78,6 +72,18 @@ The density pairs can be sourced from the same or from different data.
     - map multiple cycled attributes to a palette attribute, e.g. `[:linecolor, :markercolor] => :color`
     """
     cycle = [:color => :patchcolor]
+end
+
+function attribute_groups(::Type{<:Violin})
+    groups = default_attribute_groups()
+    push!(
+        groups, "Median Line Attributes" => sort!(
+            [
+                :mediancolor, :medianlinewidth, :show_median,
+            ]
+        )
+    )
+    return groups
 end
 
 conversion_trait(::Type{<:Violin}) = SampleBased()